+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog
 
 All Notable changes to `event-store` will be documented in this file
+
+# 1.2
+
+* Added support for upcasting to multiple events
+* Deprecated support of upcasters returning non array values